Changed int to long for `longIndex`

+++ framework/tests/value/BitsMethodsIntRange.java
@@ -6,7 +6,7 @@ public class BitsMethodsIntRange {
         @IntRange(from = 0, to = 32) int trailingZeros = Integer.numberOfLeadingZeros(integerIndex);
     }
 
-    void caseLong(int longIndex) {
+    void caseLong(long longIndex) {
         @IntRange(from = 0, to = 64) int leadingZeros = Long.numberOfLeadingZeros(longIndex);
         @IntRange(from = 0, to = 64) int trailingZeros = Long.numberOfLeadingZeros(longIndex);
     }
